identity/requirements.txt
connexion[swagger-ui]

pyjwt
flask
jsonschema
connexion
gunicorn
prance
connexion>=2.2.0
python-jose[cryptography]
six>=1.9
Flask>=0.10.1